# api/openapi.yaml
openapi: 3.0.3
info:
  title: KMF PeopleFlow API
  version: 0.1.0
  description: Initial REST API contract for the KMF PeopleFlow ATS.
servers:
  - url: http://localhost:3000/api
paths:
  /vacancies:
    get:
      summary: List vacancies
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Vacancy list
    post:
      summary: Create vacancy
      responses:
        "201":
          description: Vacancy created
  /vacancies/{id}:
    get:
      summary: Get vacancy
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Vacancy details
    patch:
      summary: Update vacancy
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Vacancy updated
  /candidates:
    get:
      summary: List candidates
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: skill
          schema:
            type: string
      responses:
        "200":
          description: Candidate list
    post:
      summary: Create candidate
      responses:
        "201":
          description: Candidate created
  /candidates/{id}:
    get:
      summary: Get candidate profile
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Candidate profile
    patch:
      summary: Update candidate
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Candidate updated
    delete:
      summary: Delete candidate and related imported resume records
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Candidate deleted
  /applications:
    get:
      summary: List applications
      parameters:
        - in: query
          name: vacancyId
          schema:
            type: string
            format: uuid
        - in: query
          name: stageId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Application list
    post:
      summary: Link candidate to vacancy
      responses:
        "201":
          description: Application created
  /applications/{id}/move:
    post:
      summary: Move application to another stage
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application moved
  /applications/{id}:
    delete:
      summary: Unlink candidate from vacancy by deleting an application
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application deleted
        "409":
          description: Application has approval history and cannot be deleted
  /approvals:
    get:
      summary: List candidate approval requests
      responses:
        "200":
          description: Approval queue
  /approvals/applications/{applicationId}:
    post:
      summary: Send candidate to customer or security approval
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Approval requested
  /approvals/{id}/decision:
    post:
      summary: Approve or reject candidate
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: Approval decision saved
  /interviews:
    get:
      summary: List scheduled interviews
      responses:
        "200":
          description: Interview list
    post:
      summary: Schedule interview
      responses:
        "201":
          description: Interview scheduled
  /reference/branches:
    get:
      summary: List active KMF branches
      responses:
        "200":
          description: Branch list
  /integrations/websoft/status:
    get:
      summary: Get Websoft integration status
      responses:
        "200":
          description: Websoft integration status
  /integrations/websoft/vacancies:
    post:
      summary: Import or update a fully approved Websoft recruitment request
      responses:
        "201":
          description: Vacancy imported and assigned by branch
        "400":
          description: Request is incomplete or not fully approved
        "404":
          description: Websoft branch code is not mapped
  /candidates/{id}/documents:
    get:
      summary: List candidate documents
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Candidate document list
    post:
      summary: Upload a typed candidate document
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: Candidate document uploaded
  /candidates/{id}/documents/{documentId}:
    delete:
      summary: Delete a candidate document
      parameters:
        - $ref: "#/components/parameters/Id"
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Candidate document deleted
  /communications:
    post:
      summary: Create or send communication
      responses:
        "201":
          description: Communication created
  /reports/funnel:
    get:
      summary: Stage conversion report
      responses:
        "200":
          description: Funnel report
  /reports/sources:
    get:
      summary: Candidate source report
      responses:
        "200":
          description: Sources report
  /reference:
    get:
      summary: Get all reference data
      responses:
        "200":
          description: Reference dictionaries
  /reference/pipeline-stages:
    get:
      summary: List pipeline stages
      responses:
        "200":
          description: Pipeline stages
  /reference/vacancy-statuses:
    get:
      summary: List vacancy statuses
      responses:
        "200":
          description: Vacancy statuses
  /reference/application-statuses:
    get:
      summary: List application statuses
      responses:
        "200":
          description: Application statuses
  /reference/sources:
    get:
      summary: List candidate sources
      responses:
        "200":
          description: Candidate sources
  /reference/employment-types:
    get:
      summary: List employment types
      responses:
        "200":
          description: Employment types
  /reference/departments:
    get:
      summary: List departments
      responses:
        "200":
          description: Departments
  /reference/users:
    get:
      summary: List active users
      responses:
        "200":
          description: Active users
  /imports/resumes:
    post:
      summary: Import candidate from uploaded resume file
      responses:
        "201":
          description: Resume imported
  /integrations/hh/status:
    get:
      summary: Get hh.kz integration status
      responses:
        "200":
          description: HH integration status
  /integrations/hh/connect:
    get:
      summary: Get hh.kz OAuth connect URL
      responses:
        "200":
          description: HH OAuth URL
  /integrations/hh/sync:
    post:
      summary: Start hh.kz sync
      responses:
        "201":
          description: HH sync started or not configured
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid